+++ src/Internal/Import-DataLayer.ps1
@@ -10,13 +10,13 @@ function Import-DataLayer {
 
     process {
         $header = Get-Content -Path ( Join-Path $Directory 'layer.yml' ) -Raw | ConvertFrom-Yaml
+        $header.Key = $Directory.Name
         $layer = New-DataLayer @header
 
         [System.IO.DirectoryInfo] $ModelDirectory = Join-Path $Directory model
         if ( $ModelDirectory.Exists ) {
             Get-ChildItem $ModelDirectory | ForEach-Object {
                 $model = Import-DataModel -Path $_
-                $model | Add-Member Key $_.BaseName -Force
                 $layer | Add-DataModel -InputObject $model
             }
         }
@@ -25,7 +25,6 @@ function Import-DataLayer {
         if ( $LayerDirectory.Exists ) {
             Get-ChildItem $LayerDirectory | ForEach-Object {
                 $sublayer = Import-DataLayer -Directory $_
-                $sublayer | Add-Member Key $_.BaseName -Force
                 $layer | Add-DataLayer -InputObject $sublayer
             }
         }
@@ -34,7 +33,6 @@ function Import-DataLayer {
         if ( $FlowDirectory.Exists ) {
             Get-ChildItem $FlowDirectory | ForEach-Object {
                 $flow = Import-DataFlow -Path $_
-                $flow | Add-Member Key $_.BaseName -Force
                 $layer | Add-DataFlow -InputObject $flow
             }
         }